requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-filter==23.2
django-phonenumber-field==7.1.0
django-rest-authtoken==2.1.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.8.11
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
